fix: address PR #119 review findings from 4th lavra-review sweep
Two genuinely new findings surfaced by an 8-agent review round against
the unchanged diff (architecture-strategist and agent-native-reviewer
independently caught the same help-text gap; pattern-recognition-specialist
caught the test-sidecar inconsistency):

- src/cli/help.rs was missing a usage line for the sessions-watch-health-check
  subcommand added in 3.6.0.
- resolve.rs and sessions_watch_legacy.rs (both new in this PR) lacked their
  own test sidecars, unlike sessions_watch_health.rs -- their tests were
  scattered into setup_tests.rs and sessions_watch_tests.rs instead. Moved
  into resolve_tests.rs / sessions_watch_legacy_tests.rs to match the
  project's documented sidecar convention.

bd syslog-mcp-8kkcn.3.4, syslog-mcp-8kkcn.3.5 (both fixed and closed).
