+++ Cargo.toml
@@ -4,7 +4,7 @@ resolver = "2"
 
 [package]
 name = "cortex"
-version = "3.6.0"
+version = "3.6.1"
 edition = "2024"
 rust-version = "1.86"
 license = "MIT"
